ai_service/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langgraph
langgraph-checkpoint-postgres
chromadb
python-dotenv
psycopg
psycopg-pool
mcp
